move deref impls to the extension file
